convert daf noncontext
