+++ src/components/donation/Steps/DonateMethods/Daf/Form.tsx
@@ -1,63 +1,31 @@
-import { yupResolver } from "@hookform/resolvers/yup";
 import CurrencySelector from "components/CurrencySelector";
-import { Field, Form as FormContainer } from "components/form";
-import { useController, useForm } from "react-hook-form";
-import { schema, stringNumber } from "schemas/shape";
-import type { DetailedCurrency } from "types/components";
+import { NativeField as Field, Form as FormContainer } from "components/form";
 import { useDonationState } from "../../Context";
 import ContinueBtn from "../../common/ContinueBtn";
 import { ProgramSelector } from "../../common/ProgramSelector";
-import { DEFAULT_PROGRAM } from "../../common/constants";
+import { usdOption } from "../../common/constants";
 import { nextFormState } from "../helpers";
-import type { FormValues as FV, Props } from "./types";
-
-/**
- * Only USD donations are permissible for DAF donations.
- * The minimum amount differs depending on which provider is selected.
- */
-const USD_CURRENCY: DetailedCurrency = { code: "usd", rate: 1, min: 1 };
+import type { Props } from "./types";
+import { useRhf } from "./useRhf";
 
 export default function Form(props: Props) {
   const { setState } = useDonationState();
-
-  const initial: FV = {
-    amount: "",
-    currency: USD_CURRENCY,
-    program: DEFAULT_PROGRAM,
-  };
-
-  const methods = useForm<FV>({
-    defaultValues: props.details || initial,
-    resolver: yupResolver(
-      schema<FV>({
-        amount: stringNumber(
-          (s) => s.required("Please enter an amount"),
-          (n) => n.positive("Amount must be greater than 0")
-        ),
-      })
-    ),
-  });
-  const { handleSubmit, control } = methods;
-
-  const { field: program } = useController<FV, "program">({
-    control: control,
-    name: "program",
-  });
+  const rhf = useRhf(props);
 
   return (
     <FormContainer
-      methods={methods}
-      onSubmit={handleSubmit((fv) =>
+      disabled={rhf.isSubmitting}
+      onSubmit={rhf.handleSubmit((fv) =>
         setState((prev) => nextFormState(prev, { ...fv, method: "daf" }))
       )}
       className="grid gap-4"
     >
       <CurrencySelector
-        currencies={[USD_CURRENCY]}
+        currencies={[usdOption]}
         label="Currency"
         // only one currency available, so can't change it
         onChange={() => {}}
-        value={USD_CURRENCY}
+        value={usdOption}
         classes={{
           label: "font-semibold",
           combobox: "field-container-donate",
@@ -65,8 +33,8 @@ export default function Form(props: Props) {
         }}
         required
       />
-      <Field<FV>
-        name="amount"
+      <Field
+        {...rhf.register("amount")}
         label="Donation amount"
         placeholder="Enter amount"
         classes={{ label: "font-semibold", container: "field-donate mt-1" }}
@@ -77,8 +45,8 @@ export default function Form(props: Props) {
       {(props.init.recipient.progDonationsAllowed ?? true) && (
         <ProgramSelector
           endowId={props.init.recipient.id}
-          program={program.value}
-          onChange={program.onChange}
+          program={rhf.program.value}
+          onChange={rhf.program.onChange}
         />
       )}
